requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.3
scikit-image==0.22.0
scikit-learn==1.4.0
tensorflow==2.15.0
google-cloud-vision==3.4.6
pillow==10.1.0
pyheif==0.7.1
pandas==2.2.0